+++ src/form-elements/FormElementBarcodeScanner.tsx
@@ -142,6 +142,7 @@ function FormElementBarcodeScanner({
                   onBlur={setIsDirty}
                   aria-describedby={ariaDescribedby}
                   autoComplete={autocompleteAttributes}
+                  aria-required={element.required}
                 />
                 <span className="ob-input-icon icon is-small is-right">
                   <MaterialIcon className="is-size-5">