+++ src/form-elements/FormElementEmail.tsx
@@ -63,6 +63,7 @@ function FormElementEmail({
               onBlur={setIsDirty}
               aria-describedby={ariaDescribedby}
               autoComplete={autocompleteAttributes}
+              aria-required={element.required}
             />
             <span className="ob-input-icon icon is-small is-right">
               <MaterialIcon className="is-size-5">email</MaterialIcon>