+++ src/form-elements/FormElementGeoscapeAddress.tsx
@@ -136,6 +136,7 @@ function FormElementGeoscapeAddress({
           setIsDirty={setIsDirty}
           aria-describedby={ariaDescribedby}
           autoComplete={autocompleteAttributes}
+          aria-required={element.required}
         />
       </FormElementLabelContainer>